move json util
