+++ src/njtai/NJTAI.java
@@ -4,6 +4,8 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.util.Enumeration;
+import java.util.Hashtable;
 import java.util.Vector;
 
 import javax.microedition.io.Connector;
@@ -18,7 +20,7 @@ import njtai.m.NJTAIM;
  * @author Feodor0090
  *
  */
-public class NJTAI implements CommandListener, ItemCommandListener {
+public class NJTAI implements CommandListener, ItemCommandListener, Runnable {
 
 	/**
 	 * Currently used URL prefix. Check {@link #getHP() home page downloading
@@ -94,11 +96,15 @@ public class NJTAI implements CommandListener, ItemCommandListener {
 	public static String[] L_PAGE = getStrings("page");
 
 	public void commandAction(Command c, Item item) {
-		
+		// TODO global command handler
 	}
 
 	public void commandAction(Command c, Displayable d) {
-		
+		// TODO global command handler
+	}
+	
+	public void run() {
+		// TODO threading
 	}
 
 	/**
@@ -460,6 +466,103 @@ public class NJTAI implements CommandListener, ItemCommandListener {
 			return getUtf(url);
 		} catch (Exception e) {
 			return null;
+		}
+	}
+	
+	// JSONUtil
+
+	/**
+	 * @param str JSON string
+	 * @return Parsed result. Hashtable, Vector or String
+	 */
+	public static Object parse(String str) {
+		int f = str.charAt(0);
+		if(f == '{') {
+			return object(str.substring(1, str.length() - 1));
+		}
+		/*if(f == '[') {
+			return array(str.substring(1, str.length() - 1));
+		}*/
+		if(f == '"') {
+			return NJTAI.replace(NJTAI.replace(str.substring(1, str.length() - 1), "\\n", "\n"), "\\\"", "\"");
+		}
+		return str;
+	}
+	
+	/**
+	 * @param str JSON object string
+	 * @return Parsed hashtable
+	 */
+	public static Hashtable object(String str) {
+		Hashtable ht = new Hashtable();
+		int unclosed = 0;
+		int index = 0;
+		int length = str.length();
+		boolean escape = false;
+		int splIndex;
+		for (; index < length; index = splIndex + 1) {
+			splIndex = index;
+			boolean quotes;
+			for (quotes = false; splIndex < length && (quotes || unclosed > 0 || str.charAt(splIndex) != ','); splIndex++) {
+				char c = str.charAt(splIndex);
+				if (!escape) {
+					if (c == '\\') {
+						escape = true;
+					}
+					if (c == '"') {
+						quotes = !quotes;
+					}
+				} else {
+					escape = false;
+				}
+				if (!quotes) {
+					if (c == '{') {
+						unclosed++;
+					} else if (c == '}') {
+						unclosed--;
+					}
+				}
+			}
+			if (quotes || unclosed > 0) {
+				throw new RuntimeException("Corrupted JSON");
+			}
+			String token = str.substring(index, splIndex);
+			int splIndex2 = token.indexOf(":");
+			ht.put(token.substring(1, splIndex2 - 1).trim(), parse(token.substring(splIndex2 + 1)));
+		}
+		return ht;
+	}
+
+	/**
+	 * @param obj Hashtable, Vector or other object to encode in JSON
+	 * @return JSON string
+	 */
+	public static String build(Object obj) {
+		if(obj instanceof Hashtable) {
+			Hashtable ht = (Hashtable) obj;
+			Enumeration en = ht.keys();
+			if (!en.hasMoreElements()) {
+				return "{}";
+			}
+			String r = "{";
+			while (true) {
+				Object e = en.nextElement();
+				r = r.concat("\"")
+				.concat(e.toString())
+				.concat("\":")
+				.concat(build(ht.get(e)));
+				if (!en.hasMoreElements()) {
+					return r.concat("}");
+				}
+				r = r.concat(",");
+			}
+		} else if(obj instanceof String) {
+			String s = (String) obj;
+			s = NJTAI.replace(s, "\n", "\\n");
+			s = NJTAI.replace(s, "\"", "\\\"");
+			return "\"".concat(s).concat("\"");
+		} else {
+			return String.valueOf(obj);
 		}
 	}
